cached_detection
When using the same dataset, this algorithm checks the last 3 folders and files in the path to determine if it is similar to the currently loaded dataset.

For example, if the dataset path is '/home/.../icosahedron/cam1/p1.png', the algorithm will only evaluate the portion '/icosahedron/cam1/p1.png' for comparison.